+++ poi-scratchpad/src/main/java/org/apache/poi/hwpf/converter/NumberFormatter.java
@@ -19,9 +19,6 @@
 
 package org.apache.poi.hwpf.converter;
 
-import java.util.Arrays;
-import java.util.Locale;
-
 import org.apache.poi.util.Beta;
 
 /**
@@ -29,75 +26,9 @@ import org.apache.poi.util.Beta;
  */
 @Beta
 public final class NumberFormatter {
-    // use char[] instead of String to speed up StringBuilder.append(), especially in JDK 11+
-    // where StringBuilder internally switched from char[] to byte[]
-    private static final char[][] ROMAN_LETTERS = Arrays.stream(
-            new String[] { "m", "cm", "d", "cd", "c", "xc", "l", "xl", "x", "ix", "v", "iv", "i" }).
-                map(String::toCharArray).
-                toArray(char[][]::new);
 
-    private static final int[] ROMAN_VALUES = { 1000, 900, 500, 400, 100, 90,
-            50, 40, 10, 9, 5, 4, 1 };
-
-    private static final int T_ARABIC = 0;
-    private static final int T_LOWER_LETTER = 4;
-    private static final int T_LOWER_ROMAN = 2;
-    private static final int T_ORDINAL = 5;
-    private static final int T_UPPER_LETTER = 3;
-    private static final int T_UPPER_ROMAN = 1;
-
+    // code was moved to org.apache.poi.util.NumberFormatter
     public static String getNumber( int num, int style ) {
-        switch ( style ) {
-            case T_UPPER_ROMAN:
-                return toRoman( num ).toUpperCase(Locale.ROOT);
-            case T_LOWER_ROMAN:
-                return toRoman( num );
-            case T_UPPER_LETTER:
-                return toLetters( num ).toUpperCase(Locale.ROOT);
-            case T_LOWER_LETTER:
-                return toLetters( num );
-            case T_ARABIC:
-            case T_ORDINAL:
-            default:
-                return String.valueOf( num );
-        }
-    }
-
-    private static String toLetters(int number) {
-        if ( number <= 0 ) {
-            throw new IllegalArgumentException( "Unsupported number: " + number );
-        }
-
-        int num = number;
-        final int radix = 26;
-
-        char[] buf = new char[33];
-        int charPos = buf.length;
-
-        while (num > 0) {
-            num--; // 1 => a, not 0 => a
-            int remainder = num % radix;
-            buf[--charPos] = (char)('a'+remainder);
-            num = (num - remainder) / radix;
-        }
-
-        return new String(buf, charPos, (buf.length - charPos));
-    }
-
-    private static String toRoman( int number ) {
-        if ( number <= 0 )
-            throw new IllegalArgumentException( "Unsupported number: " + number );
-
-        StringBuilder result = new StringBuilder();
-
-        for ( int i = 0; i < ROMAN_LETTERS.length; i++ ) {
-            char[] letter = ROMAN_LETTERS[i];
-            int value = ROMAN_VALUES[i];
-            while ( number >= value ) {
-                number -= value;
-                result.append( letter );
-            }
-        }
-        return result.toString();
+        return org.apache.poi.util.NumberFormatter.getNumber(num, style);
     }
 }
